% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.R
\name{autoplot.multiple}
\alias{autoplot.multiple}
\title{autoplot}
\usage{
\method{autoplot}{multiple}(object, ...)
}
\arguments{
\item{object}{an object of class \code{multiple}.}

\item{...}{ignored.}
}
\value{
a \code{ggplot} object
}
\description{
ggplot of the \code{multiple} objects.
}
\examples{
autoplot(multiple(2,3))
}
