% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{double}
\alias{double}
\title{double}
\usage{
double(number)
}
\arguments{
\item{number}{a numeric vector.}
}
\value{
A vector of the same length as \code{number} containing the
transformed values.
}
\description{
Double value of numbers.
}
\details{
Calculate the double values of numbers.
}
\examples{
double(2)
double(1:4)
}
