% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{fuzzydouble}
\alias{fuzzydouble}
\title{fuzzydouble}
\usage{
fuzzydouble(number, sd = 1)
}
\arguments{
\item{number}{a numeric vector.}

\item{sd}{the standard deviation of the Gaussian error added.}
}
\value{
A vector of the same length as \code{number}
containing the transformed values.
}
\description{
Double value of numbers with an error
}
\details{
Calculate the double values of numbers
and add a random error to the result.
}
\examples{
fuzzydouble(2)
fuzzydouble(1:4)
}
