% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.R, R/package.R
\docType{package}
\name{multiple}
\alias{multiple}
\title{Multiplication of a numeric vector}
\usage{
multiple(number, times = 1)
}
\arguments{
\item{number}{a numeric vector}

\item{times}{a number to multiply}
}
\value{
an object of class \code{multiple}
}
\description{
Multiples of numbers
}
\details{
This package allows simple computation of multiples
of numbers, including fast algorithms for integers \insertCite{Marcon2014c}{multiple}.
}
\examples{
multiple(1:2,3)
}
\references{
\insertAllCited{}
}
