% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.R
\name{plot.multiple}
\alias{plot.multiple}
\title{Plot objects of class multiple}
\usage{
\method{plot}{multiple}(x, y, ...)
}
\arguments{
\item{x}{a vector of numbers}

\item{y}{a vector of multiplied numbers}

\item{...}{further arguments passed to the generic method.}
}
\description{
Plot objects of class multiple
}
\examples{
plot(multiple(2,3))
}
