% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.R
\name{print.multiple}
\alias{print.multiple}
\title{Print objects of class multiple}
\usage{
\method{print}{multiple}(x, ...)
}
\arguments{
\item{x}{an object of class \code{multiple}.}

\item{...}{further arguments passed to the generic method.}
}
\description{
Print objects of class multiple
}
\examples{
print(multiple(2,3))
}
