% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{seq1_10}
\alias{seq1_10}
\title{seq1_10}
\format{
A numeric vector.
}
\source{
Values computed by the R software,
\url{https://www.r-project.org/}
}
\usage{
seq1_10
}
\description{
A sequence of numbers from 1 to 10
}
\keyword{datasets}
