% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.R
\name{summary.multiple}
\alias{summary.multiple}
\title{Summarize objects of class multiple}
\usage{
\method{summary}{multiple}(object, ...)
}
\arguments{
\item{object}{an object of class \code{multiple}.}

\item{...}{further arguments passed to the generic method.}
}
\description{
Summarize objects of class multiple
}
\examples{
summary(multiple(2,3))
}
