% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triple.R
\name{triple.integer}
\alias{triple.integer}
\title{Triple of integer vectors}
\usage{
\method{triple}{integer}(x, ...)
}
\arguments{
\item{x}{an integer vector.}

\item{...}{ignored.}
}
\value{
an integer vector.
}
\description{
This function multiplies the vector by 3.
}
\examples{
triple(1L:3L)
}
