% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triple.R
\name{triple.numeric}
\alias{triple.numeric}
\title{Triple of numeric vectors}
\usage{
\method{triple}{numeric}(x, ...)
}
\arguments{
\item{x}{a numeric vector.}

\item{...}{ignored.}
}
\value{
a numeric vector.
}
\description{
This function multiplies the vector by 3.
}
\examples{
triple(1:3)
}
