% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triple.R
\name{triple}
\alias{triple}
\title{Triple}
\usage{
triple(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
A method to multiply numbers by 3.
}
